Refactor Delete Site interface flow
Removes DeleteSiteViewController to trigger delete confirmation directly.
